% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy.R
\name{read_excel_allsheets}
\alias{read_excel_allsheets}
\title{read_excel_allsheets}
\usage{
read_excel_allsheets(filename, tibble = FALSE)
}
\arguments{
\item{filename}{filename dir}

\item{tibble}{F: tibble, T: dataframe}
}
\value{
return sheets list
}
\description{
read_excel_allsheets
}
\examples{
read_excel_allsheets("DIR.xlsx")
}
